Wrapp app in provider
This fixes the error `Error: could not find react-redux context value; please ensure the component is wrapped in a <Provider>`

Previously when it was like `<Provider store={store}/>` the `<App />` component couldn't see the Provider state as App and Provider were siblings. This PR moves App inside Provider so it's a parent child relationship.

Hopefully that makes sense! This is part of React Context: https://reactjs.org/docs/context.html might be worth a read of that. In fact I might read over it again to refresh my memory!