Add key to combine reducers arg
If you check out the format of the arguments being passed to combine reducers here: https://redux.js.org/api/combinereducers you'll notice that reducers get passed as a value in an object. That object is part of or all of your state. So if you were to log out the state in the browser you'd see it's just a single object that will grow with more and more key/value pairs as your app grows.

Hopefully that makes sense!